Moderate refactoring to allow for separate generate() callbacks for cpu and gpu (this is done to allow for a different number of Celery workers for GPU and CPU). Also fixed a bug where the session data was not being discarded upon refreshing the browser tab, added a radio button to let the user select GPU or CPU, and separated the construction of the app layout and registration of callback methods a bit more cleanly.
